requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
ldap3==2.9.1
pyasn1==0.4.8
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
